refactor: split server startup and app configuration
